refactor: constants 파일 client와 server로 분리
