+++ src/server/apis/movies.js
@@ -1,4 +1,4 @@
-import { TMDB_MOVIE_LISTS, FETCH_OPTIONS } from "../../constants.js";
+import { TMDB_MOVIE_LISTS, FETCH_OPTIONS } from "../constants.js";
 
 export const fetchMovies = async () => {
   const response = await fetch(TMDB_MOVIE_LISTS.nowPlaying, FETCH_OPTIONS);
